🚚 rename: purchase components
